Remove TCell as generic argument.
Trying to make this more actually workable. Doing gneric typing everywhere just doesn't work as I wanted it to. It annoys me, since now I have to recast to the specific type in the update functions etc. However, I don't think we can avoid this. Especially since it may be possible to have multiple cell types in a single world.